pyproject.toml
[project]
name = "twitter-user-download"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.11.3",
    "requests>=2.32.3",
    "requests-cache>=1.2.1",
    "python-dotenv>=1.1.0",
]
